Revert "split out non-windows tests"
This reverts commit e2aa7ab717cb10a9bd4c7d29b7034a70ca61878d.
